Build fix, circular imports
