+++ src/components/UIUC-Components/api-inputs/providers/AzureProviderInput.tsx
@@ -3,7 +3,10 @@ import { Text, Switch, Card, TextInput } from '@mantine/core'
 import { IconCheck, IconExternalLink, IconX } from '@tabler/icons-react'
 import { APIKeyInput } from '../LLMsApiKeyInputForm'
 import { ModelToggles } from '../ModelToggles'
-import { AzureProvider, ProviderNames } from '~/types/LLMProvider'
+import {
+  AzureProvider,
+  ProviderNames,
+} from '~/utils/modelProviders/LLMProvider'
 import { motion, AnimatePresence } from 'framer-motion'
 
 export default function AzureProviderInput({
