+++ src/components/UIUC-Components/api-inputs/providers/OpenAIProviderInput.tsx
@@ -2,7 +2,10 @@ import React from 'react'
 import { Text, Switch, Card } from '@mantine/core'
 import { IconCheck, IconExternalLink, IconX } from '@tabler/icons-react'
 import { ModelToggles } from '../ModelToggles'
-import { OpenAIProvider, ProviderNames } from '~/types/LLMProvider'
+import {
+  OpenAIProvider,
+  ProviderNames,
+} from '~/utils/modelProviders/LLMProvider'
 import { motion, AnimatePresence } from 'framer-motion'
 import { APIKeyInput } from '../LLMsApiKeyInputForm'
