+++ src/pages/api/home/home.tsx
@@ -40,8 +40,11 @@ import { type CourseMetadata } from '~/types/courseMetadata'
 import { useUser } from '@clerk/nextjs'
 import { get_user_permission } from '~/components/UIUC-Components/runAuthCheck'
 import { useRouter } from 'next/router'
-import { selectBestModel, VisionCapableModels } from '~/types/LLMProvider'
-import { OpenAIModelID } from '~/utils/modelProviders/openai'
+import {
+  selectBestModel,
+  VisionCapableModels,
+} from '~/utils/modelProviders/LLMProvider'
+import { OpenAIModelID } from '~/utils/modelProviders/types/openai'
 
 const Home = () => {
   const { t } = useTranslation('chat')
